+++ drivers/digital/encoder/digital_encoder.c
@@ -54,12 +54,12 @@ static uint32_t digital_encoder_get_status(uint8_t handle)
 	return 0;
 }
 
-static uint32_t digital_encoder_get_values(uint8_t handle,uint8_t *values,uint8_t values_number)
+static uint32_t digital_encoder_get_values(uint8_t handle,uint8_t *values,uint16_t values_number)
 {
 	return 0;
 }
 
-static uint32_t digital_encoder_set_values(uint8_t handle,uint8_t *values,uint8_t values_number)
+static uint32_t digital_encoder_set_values(uint8_t handle,uint8_t *values,uint16_t values_number)
 {
 	return 0;
 }
