+++ drivers/drivers_manager.c
@@ -111,14 +111,14 @@ ITCM_AREA_CODE uint32_t driver_start(uint32_t handle)
 	return DRIVER_REQUEST_FAILED;
 }
 
-ITCM_AREA_CODE uint32_t driver_get_values(uint32_t handle,uint8_t *values,uint8_t values_number)
+ITCM_AREA_CODE uint32_t driver_get_values(uint32_t handle,uint8_t *values,uint16_t values_number)
 {
 	if ( DriverStruct[handle]->get_values != NULL )
 		return DriverStruct[handle]->get_values(handle,values,values_number);
 	return DRIVER_REQUEST_FAILED;
 }
 
-ITCM_AREA_CODE uint32_t driver_set_values(uint32_t handle,uint8_t *values,uint8_t values_number)
+ITCM_AREA_CODE uint32_t driver_set_values(uint32_t handle,uint8_t *values,uint16_t values_number)
 {
 	if ( DriverStruct[handle]->set_values != NULL )
 		return DriverStruct[handle]->set_values(handle,values,values_number);
