+++ drivers/onchip_uart/onchip_uart.c
@@ -51,12 +51,12 @@ ITCM_AREA_CODE static uint32_t onchip_uart_get_status(uint8_t handle)
 	return 0;
 }
 
-ITCM_AREA_CODE static uint32_t onchip_uart_get_values(uint8_t handle,uint8_t *values,uint8_t values_number)
+ITCM_AREA_CODE static uint32_t onchip_uart_get_values(uint8_t handle,uint8_t *values,uint16_t values_number)
 {
 	return 0;
 }
 
-ITCM_AREA_CODE static uint32_t onchip_uart_set_values(uint8_t handle,uint8_t *values,uint8_t values_number)
+ITCM_AREA_CODE static uint32_t onchip_uart_set_values(uint8_t handle,uint8_t *values,uint16_t values_number)
 {
 	return 0;
 }
